// src/app/enum/division-type.enum.ts
/** 区划类型 */
export enum DivisionType {
  /**
   * 	无	0
   */
  None = 0,
  /**
   *	省、直辖市	1
   */
  Province = 1,
  /**
   *	市、区	2
   */
  City = 2,
  /**
   *	县、街道	3
   */
  County = 3,
  /**
   *	居委会	4
   */
  Committees = 4,
  /**
   *	小区、社区	5
   */
  Community = 5,
}